refactor: move public type aliases to `types.py`, rename `_types.py` to `_internal_types.py`
