+++ src/apify_client/_internal_types.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from typing import Literal
+
+TerminalActorJobStatus = Literal['SUCCEEDED', 'FAILED', 'TIMED-OUT', 'ABORTED']
+"""Subset of `ActorJobStatus` values that indicate the job has finished and will not change again."""
