+++ src/apify_client/_resource_clients/actor_collection.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     from datetime import timedelta
 
     from apify_client._models import ActorShort
-    from apify_client._types import Timeout
+    from apify_client.types import Timeout
 
 
 @docs_group('Resource clients')
