+++ src/apify_client/_resource_clients/dataset.py
@@ -21,8 +21,8 @@ if TYPE_CHECKING:
     from datetime import timedelta
 
     from apify_client._literals import GeneralAccess
-    from apify_client._types import JsonSerializable, Timeout
     from apify_client.http_clients import HttpResponse
+    from apify_client.types import JsonSerializable, Timeout
 
 
 @docs_group('Other')
