+++ src/apify_client/_resource_clients/dataset_collection.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
 
     from apify_client._literals import StorageOwnership
     from apify_client._models import DatasetListItem
-    from apify_client._types import Timeout
+    from apify_client.types import Timeout
 
 
 @docs_group('Resource clients')
