+++ src/apify_client/_resource_clients/request_queue.py
@@ -46,7 +46,7 @@ if TYPE_CHECKING:
 
     from apify_client._literals import GeneralAccess
     from apify_client._typeddicts import RequestDict, RequestDraftDeleteDict, RequestDraftDict
-    from apify_client._types import Timeout
+    from apify_client.types import Timeout
 
 _RQ_MAX_REQUESTS_PER_BATCH = 25
 _MAX_PAYLOAD_SIZE_BYTES = 9 * 1024 * 1024  # 9 MB
