+++ src/apify_client/_resource_clients/webhook_collection.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
 
     from apify_client._literals import WebhookEventType
     from apify_client._models import Webhook, WebhookShort
-    from apify_client._types import Timeout
+    from apify_client.types import Timeout
 
 
 @docs_group('Resource clients')
